획원가입, 댓글기능 완료

+++ pages/api/post/delete.js
@@ -3,12 +3,21 @@ import { ObjectId } from 'mongodb'
 
 export default async function handler(요청, 응답) {
   if (요청.method == 'DELETE') {
-    let db = (await connectDB).db('forum')
-    let result = await db
+    console.log(요청.body)
+    let session = await getServerSession(요청, 응답, authOptions)
+
+    const db = (await connectDB).db('forum')
+    let 찾은거 = await db
       .collection('post')
-      .deleteOne({ _id: new ObjectId(요청.body) })
-    console.log(result)
-    응답.status(200).json('삭제완료')
-    응답.redirect(302, '/list')
+      .findOne({ _id: new ObjectId(요청.body) })
+
+    if (찾은거.author == session.user.email) {
+      let result = await db
+        .collection('post')
+        .deleteOne({ _id: new ObjectId(요청.body) })
+      return 응답.status(200).json('삭제완료')
+    } else {
+      return 응답.status(500).json('현재유저와 작성자 불일치')
+    }
   }
 }
